+++ example/lib/main.dart
@@ -1,7 +1,4 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:flutter/services.dart';
 import 'package:zebra_scanner_plugin/zebra_scanner_plugin.dart';
 
 void main() {
@@ -16,45 +13,53 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String _platformVersion = 'Unknown';
+  List<String> scannedCodes = [];
 
-  @override
-  void initState() {
-    super.initState();
-    initPlatformState();
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
-    String platformVersion;
-    // Platform messages may fail, so we use a try/catch PlatformException.
-    // We also handle the message potentially returning null.
-    try {
-      platformVersion =
-          await ZebraScannerPlugin.platformVersion ?? 'Unknown platform version';
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
-    if (!mounted) return;
-
-    setState(() {
-      _platformVersion = platformVersion;
-    });
-  }
+  ScrollController listScrollController = ScrollController();
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('Zebra Scanner Plugin example app'),
+          actions: [
+            PopupMenuButton(itemBuilder: (context) => [
+              PopupMenuItem(
+                child: const Text("Scan and connect"),
+                onTap: () async{
+                  await ZebraScannerPlugin.connect;
+                },
+                value: 1,
+              ),
+            ])
+          ],
         ),
-        body: Center(
-          child: Text('Running on: $_platformVersion\n'),
+        body: StreamBuilder<String>(
+            stream: ZebraScannerPlugin.barcodeStream,
+            builder: (context, snapshot) {
+              String? data;
+              if (snapshot.hasData) {
+                data = snapshot.data;
+                scannedCodes.add(data!);
+                if (listScrollController.hasClients) {
+                  final position = listScrollController.position.maxScrollExtent;
+                  listScrollController.jumpTo(position);
+                }
+              } else if (snapshot.hasError){
+                print(snapshot.error);
+              }
+              return Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: ListView.builder(itemBuilder: (context, index) {
+                  return ListTile(title: Text(scannedCodes[index]),
+                      leading: const Icon(Icons.qr_code,));
+                },
+                  controller: listScrollController,
+                  itemCount: scannedCodes.length,
+                ),
+              );
+            }
         ),
       ),
     );