+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 0.0.1
 
-# initial commit
+## initial commit
 
 ## 0.0.2
 
- # Created src directory and exported the file through zebra_scanner_plugin file
+## Created src directory and exported the file through zebra_scanner_plugin file
+
+## 0.0.3
+
+## Integrated zebra sdk for android and added barcode stream